slight refactor, test seed
