+++ front-end/src/Views/BlogPage/components/Filter/Filter.jsx
@@ -7,12 +7,14 @@ import { UserContext } from "context";
 import { IndexLinkContainer } from "react-router-bootstrap";
 import { staffTypes } from "constants/staffTypes";
 import { userTypes } from "constants/userTypes";
+import NewVideoBlogModal from "components/NewVideoBlogModal/NewVideoBlogModal";
 
 const Filter = ({ setFilter, filtersBlog, count }) => {
   const [filters, setFilters] = useState({ categories: [], tags: [] });
   const [category, setCategory] = useState(DEFAULT_INPUT);
   const [tag, setTag] = useState(DEFAULT_INPUT);
   const [err, setErr] = useState(false);
+  const [showModal, setShowModal] = useState(false);
   const user = useContext(UserContext);
 
   useEffect(() => {
@@ -61,59 +63,67 @@ const Filter = ({ setFilter, filtersBlog, count }) => {
   ) : null;
 
   return (
-    <Form className="ml-3 mr-3 mb-3" onSubmit={filter}>
-      <Form.Row>
-        <FormGroup
-          as={Col}
-          xs={12}
-          md={6}
-          header="Kategoria"
-          array={filters.categories}
-          val={category}
-          setVal={setCategory}
-          type="select"
-          id="category"
-        />
-        <FormGroup
-          as={Col}
-          xs={12}
-          md={6}
-          header="Tag"
-          array={filters.tags}
-          val={tag}
-          setVal={setTag}
-          type="select"
-          id="tag"
-        />
-      </Form.Row>
-      <div>
-        <Button type="submit" className="mr-3" variant="primary">
-          Filtruj posty
-        </Button>
-        <Button
-          variant="outline-primary"
-          className="mr-3"
-          onClick={clearFilter}
-        >
-          Wyczyść filtry
-        </Button>
-      </div>
-      {msg}
-      {user &&
-      user.type === userTypes.STAFF &&
-      user.data.group_type.includes(staffTypes.BLOG_CREATOR) ? (
-        <IndexLinkContainer as={Button} to="/blog/newPost">
-          <Button variant="success" className="mt-2">
-            Stwórz nowy post
+    <>
+      <Form className="ml-3 mr-3 mb-3" onSubmit={filter}>
+        <Form.Row>
+          <FormGroup
+              as={Col}
+              xs={12}
+              md={6}
+              header="Kategoria"
+              array={filters.categories}
+              val={category}
+              setVal={setCategory}
+              type="select"
+              id="category"
+          />
+          <FormGroup
+              as={Col}
+              xs={12}
+              md={6}
+              header="Tag"
+              array={filters.tags}
+              val={tag}
+              setVal={setTag}
+              type="select"
+              id="tag"
+          />
+        </Form.Row>
+        <div>
+          <Button type="submit" className="mr-3" variant="primary">
+            Filtruj posty
           </Button>
-        </IndexLinkContainer>
-      ) : null}
-      <div>
-        {count !== 0 && (
-          <small className="blog__countText">{`Ilość znalezionych postów: ${count}`}</small>
-        )}
-      </div>
-    </Form>
+          <Button
+              variant="outline-primary"
+              className="mr-3"
+              onClick={clearFilter}
+          >
+            Wyczyść filtry
+          </Button>
+        </div>
+        {msg}
+        {user &&
+        user.type === userTypes.STAFF &&
+        user.data.group_type.includes(staffTypes.BLOG_CREATOR) ? (
+            <>
+              <IndexLinkContainer as={Button} to="/blog/newPost">
+                <Button variant="success" className="mt-2">
+                  Stwórz nowy post
+                </Button>
+              </IndexLinkContainer>
+              <Button variant="success" className="mt-2 mx-3" onClick={() => setShowModal(true)}>
+                Stwórz nowy wideoblog
+              </Button>
+            </>
+        ) : null}
+        <div>
+          {count !== 0 && (
+              <small className="blog__countText">{`Ilość znalezionych postów: ${count}`}</small>
+          )}
+        </div>
+      </Form>
+      <NewVideoBlogModal show={showModal} setShow={setShowModal} user={user} />
+    </>
   );
 };
